Factor out letter-content-util.tsx.
