+++ frontend/lib/norent/tests/letter-content.test.tsx
@@ -3,7 +3,6 @@ import ReactTestingLibraryPal from "../../tests/rtl-pal";
 import {
   NorentLetterContent,
   noRentSampleLetterProps,
-  getStreetWithApt,
 } from "../letter-content";
 import { initNationalMetadataForTesting } from "../letter-builder/tests/national-metadata-test-util";
 import { override } from "../../tests/util";
@@ -24,19 +23,5 @@ describe("<NorentLetterContent>", () => {
       )
     );
     expect(pal.rr.container).toMatchSnapshot();
-  });
-});
-
-describe("getStreetWithApt()", () => {
-  it("returns only street if apt is blank", () => {
-    expect(getStreetWithApt({ street: "1234 Boop Way", aptNumber: "" })).toBe(
-      "1234 Boop Way"
-    );
-  });
-
-  it("returns street w/ apt if apt is present", () => {
-    expect(getStreetWithApt({ street: "1234 Boop Way", aptNumber: "2" })).toBe(
-      "1234 Boop Way #2"
-    );
   });
 });
